cert-config/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
bit==0.8.0
bitarray==2.9.2
cachetools==5.3.1
cbor2==5.4.6
cert-core==3.0.0
cert-issuer==3.5.0
cert-schema==3.4.2
cert-tools==3.0.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
clickclick==20.10.2
coincurve==18.0.0
colorama==0.4.6
ConfigArgParse==0.13.0
connexion==2.14.2
cytoolz==0.12.2
decorator==5.1.1
distlib==0.3.7
Django==4.2.6
dnspython==2.4.2
ecdsa==0.18.0
eth-account==0.10.0
eth-hash==0.6.0
eth-keyfile==0.7.0
eth-keys==0.5.0
eth-rlp==1.0.0
eth-typing==4.0.0
eth-utils==3.0.0
eth_abi==5.0.0
filelock==3.12.4
Flask==2.2.5
Flask-PyMongo==2.3.0
frozenlist==1.4.1
glob2==0.6
hexbytes==0.3.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
inflection==0.5.1
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpath-rw==1.4.0
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
lds-merkle-proof-2019==0.1.0
lru-dict==1.2.0
MarkupSafe==2.1.3
merkletools==1.0.3
mock==2.0.0
morphys==1.0
multiaddr==0.0.9
multidict==6.0.4
mysql-connector-python==8.2.0
mysqlclient==2.2.0
netaddr==0.10.1
packaging==23.2
parsimonious==0.9.0
pbr==5.11.1
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
pluggy==1.3.0
ply==3.11
protobuf==4.25.2
py-multibase==1.0.3
pycoin==0.80
pycparser==2.21
pycryptodome==3.20.0
PyLD==1.0.5
pymongo==4.5.0
pyproject-api==1.6.1
pyrsistent==0.20.0
pysha3==1.0.2
python-baseconv==1.2.2
python-bitcoinlib==0.12.2
python-dateutil==2.8.2
pytz==2023.3.post1
pyunormalize==15.1.0
PyYAML==6.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rlp==4.0.0
rpds-py==0.17.1
sha256==1.0
simplekv==0.14.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
toolz==0.12.0
tox==4.11.3
typing_extensions==4.8.0
urllib3==2.0.6
validators==0.22.0
varint==1.0.2
virtualenv==20.24.5
websockets==12.0
Werkzeug==2.2.3
yarl==1.9.4
zipp==3.17.0
